pyproject.toml
[project]
name = "rmgx-rag-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.23",
    "chromadb>=1.0.20",
    "fastapi>=0.116.1",
    "from-root>=1.3.0",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langchain-pinecone>=0.2.12",
    "numpy>=2.3.2",
    "pinecone-client>=6.0.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.0",
    "uvicorn[standard]>=0.35.0",
]
